build: reorganización de los archivos de la librería
