+++ Etapa_de_Control/lib/gravicap/cap_motor.h
@@ -1,0 +1,13 @@
+#ifndef _cap_motor_h
+#define _cap_motor_h
+
+#include <stdio.h>
+#include <string.h>
+#include <math.h>
+#include "pico/stdlib.h"
+
+void carga_motor();
+int descarga_motor();
+void motor_stop();
+
+#endif // _cap_motor_h
